+++ modules/customeFormats/index.ts
@@ -2,3 +2,4 @@ export { default as DividerBlot } from './divider';
 export { default as Image } from './image';
 export { default as ListItem } from './listItem';
 export { default as QSyntax } from './syntax';
+export { default as VideoBlot } from './video';